Function poker::play

pub fn play(player1: String, player2: String, table: String) -> String

play Initializes a game with 2 players with 2 cards each one and 5 community cards. It will try to find the highest hand for every player. Returns number of the player with the highest hand, the name of the hand and the cards.

Example: INPUT OUTPUT 8D 7C 1 HIGH_CARD K9875 7D 6C KS 9D 5C 3S 2D