poker

Module poker

Entry point for the poker lib.

Modules

autoactor

betting_strategy

card

compare

game

names

player

sequence

Functions

new_game_one_player

Create a new game with one supplied player and the supplied number of auto players. Supply an interactive player to create a one player game. The auto players use either the six_max or modest betting strategies.

new_game_with_players

Create a new game with the supplied players

rotate_vector

Rotate a vector (V) by a given index (I). The rotation moves the elements starting from V[I] to the front, followed by the elements V[..I].