poker::poker::betting_strategy

Function six_max

pub fn six_max(args: BetArgs, hole_cards: (Card, Card), bank_roll: usize) -> Bet

A strategy that folds at the preflop for hands not in the top 15% of pairs of cards. If we do have a good pair of hole cards, then raise twice in each betting stage, so as we can afford it.