poker::poker

Function new_game_one_player

pub fn new_game_one_player(
    player: Player,
    big_blind: usize,
    num_auto_players: u8,
) -> Game

Create a new game with one supplied player and the supplied number of auto players. Supply an interactive player to create a one player game. The auto players use either the six_max or modest betting strategies.