poker::poker::player

Struct Player

pub struct Player {
    pub name: String,
    pub hole: Option<(Card, Card)>,
    pub bet: usize,
    pub bank_roll: usize,
    pub all_in: bool,
    pub folded: bool,
    pub actor: Box<dyn Actor + 'static>,
}

The Player struct.

Fields
name: Stringhole: Option<(Card, Card)>bet: usizebank_roll: usizeall_in: boolfolded: boolactor: Box<dyn Actor + 'static>
Implementations

impl Player

Implementation of Player.

pub fn build(name: &str, actor: impl Actor + 'static) -> Player

Construct a Player instance with the supplied actor.

pub fn set_name_and_bank_roll(&mut self, name: &String, bank_roll: usize)

Set name and bank roll at the beginning of a game. Needed because the name might need to be changed to become unique, and so that this info can be passed to remote clients.

pub fn hole_cards(&mut self, (h1, h2): (Card, Card))

Accept the hole cards and pass them on to the actor.

pub fn place_bet(&mut self, args: BetArgs) -> Option<Bet>

Place a bet by asking the actor to do it.

pub fn update(&self, msg: &Msg)

Respond to an incoming message by asking the actor to do it.

pub fn ante_up(&mut self, blind: usize) -> Option<usize>

Pay the required amount to join a round.