poker::poker::player

Trait Actor

pub trait Actor: Debug {
    // Required methods
    fn set_name_and_bank_roll(&self, name: &String, bank_roll: usize);
    fn hole_cards(&self, hole_cards: (Card, Card));
    fn place_bet(
        &mut self,
        args: BetArgs,
        hole_cards: (Card, Card),
        bank_roll: usize,
    ) -> Option<Bet>;
    fn update(&self, msg: &Msg);
}

The Actor trait is the component part of players which places bets and responds to messages.

Required Methods

fn set_name_and_bank_roll(&self, name: &String, bank_roll: usize)

Accept the name and bank roll at the beginning of the game.

fn hole_cards(&self, hole_cards: (Card, Card))

Accept the hole cards at the beginning of a round.

fn place_bet( &mut self, args: BetArgs, hole_cards: (Card, Card), bank_roll: usize, ) -> Option<Bet>

Place a bet.

fn update(&self, msg: &Msg)

Receive an update message, e.g. the status of the game or information about the winner of a round or game.